Day5 new solution using binary representation for char sequence

+++ src/day5/part1.rs
@@ -1,6 +1,6 @@
 use crate::day5::common::{parse_test, parse_final};
 
-pub fn exo() -> u16 {
+pub fn exo() -> u32 {
     let boardingPasses = parse_final();
     let mut max = 0;
     for boardingpass in boardingPasses {
